requirements.txt
colorama==0.4.6
geomloss==0.2.6
matplotlib==3.4.1
numpy==1.26.4
opencv_python==4.9.0.80
pandas==1.2.4
Pillow==10.2.0
scikit_learn==1.3.2
scipy==1.12.0
skimage==0.0
torch==2.1.2
torchvision==0.16.2
Wand==0.6.13
